+++ Firebase/Core/Public/FIRConfiguration.h
