+++ Firebase/Core/Public/FirebaseCore.h
